+++ web/test/value-graph.test.ts
@@ -38,6 +38,8 @@ function holding(over: Partial<HoldingView>): HoldingView {
     todayMovePct: null,
     todayMoveIsStale: false,
     todayFxMoveEur: null,
+    priorCloseValueEur: null,
+    priorCloseValueUsd: null,
     weight: null,
     unrealisedPlEur: null,
     totalGrowthPct: null,